Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -2,9 +2,9 @@
 
 int main()
 {
-    float pi=3.14;
+    int pi=3.14;
     int r;
     scanf("%d",&r);
-    printf("Area: %.2f",r*r*pi);
+    printf("Area: %.2f",(float)r*r*pi);
     return 0;
 }
